dv360_parser/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-bigquery==3.4.0
requests==2.28.1
pandas==1.5.1
google-auth-oauthlib==0.7.1
oauth2client==4.1.3
oauthlib==3.2.2
google-api-python-client==2.66.0
dask